backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.8.0
python-engineio==4.7.1
python-multipart==0.0.6
apscheduler==3.10.4
python-dotenv==1.0.0
aiofiles==23.2.1
